Consistent download path
Prior to this commit the downloaded tarball would always be downloaded
to the directory from which the op5-lxd script was called from.

Now the tarballs are either saved to the directory of the script itself,
or alternatively it can be configured with the `OP5LXD_TARBALL_PATH`
environment variable.